+++ packages/apps-config/src/api/spec/contracts.ts
