+++ packages/apps-config/src/ui/colors.ts
@@ -110,7 +110,7 @@ const chainRoccoMandala = '#173DC9';
 const chainRoccoAres = '#70FF8B';
 const chainAresGladios = '#1295F0';
 const chainRoccoBifrost = '#5a25f0';
-const chainRococoCanvas = '#000000';
+const chainRococoContracts = '#000000';
 const chainRoccoDarwinia = 'linear-gradient(-45deg, #FE3876 0%, #7C30DD 71%, #3A30DD 100%)';
 const chainRoccoDataHighway = 'linear-gradient(-90deg, #9400D3 0%, #5A5CA9 50%, #00BFFF 100%)';
 const chainRococoEave = '#900048';
@@ -265,7 +265,7 @@ export const chainColors: Record<string, string> = Object.entries({
   'Calamari Parachain Development': chainCalamari,
   'Calamari Parachain Local': chainCalamari,
   'Calamari Parachain Testnet': chainCalamari,
-  'Canvas on Rococo': chainRococoCanvas,
+  'Contracts on Rococo': chainRococoContracts,
   Centrifuge: nodeCentrifuge,
   ChainX: chainChainx,
   Clover: chainClover,
