+++ packages/apps/public/locales/fr/translation.json
@@ -1328,7 +1328,7 @@
   "rpc.pontem": "Pontem",
   "rpc.rococo": "Rococo",
   "rpc.rococo.acala": "Mandala PC1",
-  "rpc.rococo.canvas": "Canvas",
+  "rpc.rococo.contracts": "Contracts",
   "rpc.rococo.clover": "Clover PC1",
   "rpc.rococo.crust": "Crust PC1",
   "rpc.rococo.darwinia": "Darwinia PC2",
