+++ packages/apps/public/locales/it/translation.json
@@ -1804,9 +1804,9 @@
   "rpc.rococo.bajun": "",
   "rpc.rococo.bifrost": "",
   "rpc.rococo.bitcountry": "",
-  "rpc.rococo.canvas": "",
   "rpc.rococo.chainx": "",
   "rpc.rococo.clover": "",
+  "rpc.rococo.contracts": "",
   "rpc.rococo.crab": "",
   "rpc.rococo.crust": "",
   "rpc.rococo.darwinia": "",
