+++ packages/apps/public/locales/ko/translation.json
@@ -1290,7 +1290,7 @@
   "rpc.pontem": "다리",
   "rpc.rococo": "로코코",
   "rpc.rococo.acala": "만달라 PC1",
-  "rpc.rococo.canvas": "캔버스",
+  "rpc.rococo.contracts": "캔버스",
   "rpc.rococo.darwinia": "다위니아 PC2",
   "rpc.rococo.laminar": "격동 PC1",
   "rpc.rococo.plasm": "플라즈마 PC1",
