+++ packages/apps/public/locales/ru/translation.json
@@ -1872,9 +1872,9 @@
   "rpc.rococo.bajun": "",
   "rpc.rococo.bifrost": "",
   "rpc.rococo.bitcountry": "",
-  "rpc.rococo.canvas": "",
   "rpc.rococo.chainx": "",
   "rpc.rococo.clover": "",
+  "rpc.rococo.contracts": "",
   "rpc.rococo.crab": "",
   "rpc.rococo.crust": "",
   "rpc.rococo.darwinia": "",
